Moving blog data change verification to separate file in order to remove non private methods
