+++ src/main/java/com/jvm_bloggers/core/data_fetching/blogs/BloggerChangedVerifier.java
@@ -1,0 +1,28 @@
+package com.jvm_bloggers.core.data_fetching.blogs;
+
+import com.jvm_bloggers.core.data_fetching.blogs.domain.Blog;
+import com.jvm_bloggers.core.data_fetching.blogs.json_data.BloggerEntry;
+
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.stereotype.Component;
+
+import java.util.Objects;
+
+@Component
+public class BloggerChangedVerifier {
+
+    public boolean pendingChanges(Blog blog, BloggerEntry bloggerEntry) {
+        return !Objects.equals(blog.getAuthor(), bloggerEntry.getName())
+            || !Objects.equals(blog.getJsonId(), bloggerEntry.getJsonId())
+            || !StringUtils.equalsIgnoreCase(blog.getRss(), bloggerEntry.getRss())
+            || !Objects.equals(blog.getBlogType(), bloggerEntry.getBlogType())
+            || !Objects.equals(blog.getTwitter(), bloggerEntry.getTwitter())
+            || urlFromRssIsValidAndDifferentThanExistingOne(blog, bloggerEntry);
+    }
+
+    private boolean urlFromRssIsValidAndDifferentThanExistingOne(Blog blog,
+                                                                 BloggerEntry bloggerEntry) {
+        return StringUtils.isNotBlank(bloggerEntry.getUrl())
+            && !StringUtils.equalsIgnoreCase(blog.getUrl(), bloggerEntry.getUrl());
+    }
+}
